+++ src/append/testing.rs
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use log::Record;
-
 use crate::Diagnostic;
+use crate::Error;
 use crate::Layout;
 use crate::append::Append;
 use crate::layout::TextLayout;
@@ -60,7 +59,11 @@ impl Testing {
 }
 
 impl Append for Testing {
-    fn append(&self, record: &Record, diagnostics: &[Box<dyn Diagnostic>]) -> anyhow::Result<()> {
+    fn append(
+        &self,
+        record: &log::Record,
+        diagnostics: &[Box<dyn Diagnostic>],
+    ) -> Result<(), Error> {
         let bytes = self.layout.format(record, diagnostics)?;
         eprintln!("{}", String::from_utf8_lossy(&bytes));
         Ok(())